backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
fastapi-sso>=0.15.0
slowapi>=0.1.9

